Create 16_SEQ2.ino
Version 2 of the CV Gate Sequencer.  There was a bug, the notes would chirp due to having a millisecond of the old note before the gate was closed and opened for the next note.  Now fixed.

Code was completely re-written and optimized, runs much faster now and is MUCH cleaner to understand.  Same functionality.