bug fix
There was a bug, when using Internal source, when the clock triggered a chord change, the chord would Chirp.  This was due to the previous chord being played for about 30ms before changing the chord.  Delay was caused by memory dump to oled, which blocked the pwm frequency change.  I moved the timing/location of the drawline oled update and solved the issue.